renamed file inside search removed-'_search_'.  Removed  ip & coordinates from other file.  Updated mocked_hits.  status code test using requests.codes.ok. Changed page to string variable.
